examples/servers/sns-sqs/function/requirements.txt
awslabs.amazon-sns-sqs-mcp-server==2.0.2

# For testing, this module is installed from local files.
# Uncomment this line to build using the module from PyPi
# run-mcp-servers-with-aws-lambda >= 0.0.1
